Change to new Arduino-Makefile
Change to the new Makefile option to handle more consistent development
outside of the Arduino IDE.

The Makefile change also has the side effect of forcing the moving of
the formally duplicated code in the old util/ sub-dir, into their own
libraries inside the ..../sketchbook/libraries location. All these utils
now finally have their own libraries which will still be placed under
version control at some point in the future.

Due to the use of the DEBUG macro in one of the libraries used, this
macro can not be used with this name inside this (or future) projects
for local debugging. For this reason the new _DEBUG macro is used (and
updated in all sources affected), to set debug mode.

Also fixed some linting issues highlighted by using the -Werror flag,
both in the local project code, as well as some of the external
libraries.